replace time.Now() clock with a mockable clock from jonboulle/clockwork. Covnert float64 amount fields to int64 to standardize with blockchain and other financial sysetms and avoid rounding errors. Combine dunning workers into a single worker for simplicity. Remove get-access-status since it is redundnat and stupid.

+++ internal/db/repo/solana_transaction.go
@@ -13,7 +13,7 @@ type SolanaTransactionRepo struct {
 
 func NewSolanaTransactionRepo(d *db.DB) *SolanaTransactionRepo { return &SolanaTransactionRepo{db: d} }
 
-func (r *SolanaTransactionRepo) MarkConfirmedByUserAndAmount(ctx context.Context, userID string, amount float64, signature string) error {
+func (r *SolanaTransactionRepo) MarkConfirmedByUserAndAmount(ctx context.Context, userID string, amount int64, signature string) error {
 	_, err := r.db.GetDB().NewUpdate().
 		Model((*models.SolanaTransaction)(nil)).
 		Set("status = ?", "confirmed").
